Move Location row button styles to separate file
